Database schema updated.

+++ Domain/EShop.cs
@@ -4,7 +4,7 @@ namespace Domain
 {
     public class EShop
     {
-        public double Id { get; set; }
+        public long Id { get; set; }
         public string Title { get; set; }
         public string Url { get; set; }
