+++ Domain/Product.cs
@@ -4,9 +4,9 @@ namespace Domain
 {
     public class Product
     {
-        public double Id { get; set; }
+        public long Id { get; set; }
         public string Title { get; set; }
-        public double Popularity { get; set; }
+        public long Popularity { get; set; }
 
         public ICollection<RealProduct> RealProducts { get; set; }
         public ICollection<ProductAttribute> Attributes { get; set; }
