+++ Domain/ProductAttribute.cs
@@ -2,10 +2,10 @@
 {
     public class ProductAttribute
     {
-        public double ProductId { get; set; }
+        public long ProductId { get; set; }
         public Product Product { get; set; }
 
-        public double AttributeId { get; set; }
+        public long AttributeId { get; set; }
         public Attribute Attribute { get; set; }
 
     }
